requirements.txt
# Dev environment
pip
autopep8
numpy
opencv-python-headless
tqdm
pillow
opencv-python

# App
flask